# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Api
  version: 0.2.0
  description: L00T.tv MVP API
servers:
  - url: /api
...
paths:
...
  /auth/nonce:
    post:
      operationId: authNonce
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WalletRequest" }
      responses:
        "200":
          description: Nonce and SIWE-style message
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NonceResponse" }
  /auth/verify:
    post:
      operationId: authVerify
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthVerifyRequest" }
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
...
  /me:
    get:
      operationId: getMe
      tags: [users]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
    patch:
      operationId: updateMe
      tags: [users]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfileUpdate" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
...
  /creators:
    get:
      operationId: getCreators
      tags: [creators]
      parameters:
        - { name: q, in: query, schema: { type: string } }
      responses:
        "200":
          description: Creators
          content:
            application/json:
              schema:
                type: object
                properties:
                  creators:
                    type: array
                    items: { $ref: "#/components/schemas/CreatorProfile" }
                required: [creators]
    post:
      operationId: createCreator
      tags: [creators]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatorProfileInput" }
      responses:
        "201":
          description: Created creator
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatorResponse" }
  /creators/me:
    patch:
      operationId: updateCreatorMe
      tags: [creators]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatorProfileInput" }
      responses:
        "200":
          description: Updated creator
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatorResponse" }
  /creators/{username}:
    get:
      operationId: getCreator
      tags: [creators]
      parameters:
        - { $ref: "#/components/parameters/Username" }
      responses:
        "200":
          description: Creator
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatorDetailResponse" }
...
  /creators/{username}/posts:
    get:
      operationId: getCreatorPosts
      tags: [posts]
      parameters:
        - { $ref: "#/components/parameters/Username" }
      responses:
        "200":
          description: Creator posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items: { $ref: "#/components/schemas/Post" }
                  subscribed: { type: boolean }
                required: [posts, subscribed]
  /streams:
    get:
      operationId: getStreams
      tags: [streams]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [live] } }
      responses:
        "200":
          description: Streams
          content:
            application/json:
              schema:
                type: object
                properties:
                  streams:
                    type: array
                    items: { $ref: "#/components/schemas/Stream" }
                required: [streams]
    post:
      operationId: createStream
      tags: [streams]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StreamInput" }
      responses:
        "201":
          description: Created stream
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StreamResponse" }
  /streams/{id}:
    get:
      operationId: getStream
      tags: [streams]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "200":
          description: Stream
          content:
            application/json:
              schema:
                type: object
                properties:
                  stream: { $ref: "#/components/schemas/Stream" }
                  locked: { type: boolean }
                required: [stream, locked]
    patch:
      operationId: updateStream
      tags: [streams]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StreamInput" }
      responses:
        "200":
          description: Updated stream
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StreamResponse" }
  /streams/{id}/start:
    post:
      operationId: startStream
      tags: [streams]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "200":
          description: Started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StreamResponse" }
  /streams/{id}/end:
    post:
      operationId: endStream
      tags: [streams]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "200":
          description: Ended
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StreamResponse" }
  /streams/{id}/regenerate-key:
    post:
      operationId: regenerateStreamKey
      tags: [streams]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "200":
          description: Regenerated key
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StreamResponse" }
  /streams/{id}/chat:
    get:
      operationId: getStreamChat
      tags: [chat]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items: { $ref: "#/components/schemas/ChatMessage" }
                required: [messages]
    post:
      operationId: sendStreamChat
      tags: [chat]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message: { type: string }
              required: [message]
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { $ref: "#/components/schemas/ChatMessage" }
                required: [message]
  /posts:
    post:
      operationId: createPost
      tags: [posts]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostInput" }
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                type: object
                properties:
                  post: { $ref: "#/components/schemas/Post" }
                required: [post]
  /posts/{id}:
    patch:
      operationId: updatePost
      tags: [posts]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostInput" }
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                type: object
                properties:
                  post: { $ref: "#/components/schemas/Post" }
                required: [post]
    delete:
      operationId: deletePost
      tags: [posts]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /donations/verify:
    post:
      operationId: verifyDonation
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DonationVerifyRequest" }
      responses:
        "201":
          description: Verified donation
          content:
            application/json:
              schema:
                type: object
                properties:
                  donation: { $ref: "#/components/schemas/Donation" }
                required: [donation]
  /subscriptions/verify:
    post:
      operationId: verifySubscription
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubscriptionVerifyRequest" }
      responses:
        "201":
          description: Verified subscription
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscription: { $ref: "#/components/schemas/Subscription" }
                required: [subscription]
...
  /creators/{username}/subscription-status:
    get:
      operationId: getSubscriptionStatus
      tags: [payments]
      parameters:
        - { $ref: "#/components/parameters/Username" }
      responses:
        "200":
          description: Subscription status
          content:
            application/json:
              schema:
                type: object
                properties:
                  active: { type: boolean }
                  plan: { $ref: "#/components/schemas/SubscriptionPlan" }
                required: [active]
  /creators/{username}/donations:
    get:
      operationId: getCreatorDonations
      tags: [payments]
      parameters:
        - { $ref: "#/components/parameters/Username" }
      responses:
        "200":
          description: Creator donations
          content:
            application/json:
              schema:
                type: object
                properties:
                  donations:
                    type: array
                    items: { $ref: "#/components/schemas/Donation" }
                required: [donations]
  /streams/{id}/donations:
    get:
      operationId: getStreamDonations
      tags: [payments]
      parameters:
        - { $ref: "#/components/parameters/Id" }
      responses:
        "200":
          description: Stream donations
          content:
            application/json:
              schema:
                type: object
                properties:
                  donations:
                    type: array
                    items: { $ref: "#/components/schemas/Donation" }
                required: [donations]
...
components:
  parameters:
...
    Username:
      name: username
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    WalletRequest:
      type: object
      properties:
        walletAddress: { type: string }
      required: [walletAddress]
    AuthVerifyRequest:
      type: object
      properties:
        walletAddress: { type: string }
        signature: { type: string }
      required: [walletAddress, signature]
    NonceResponse:
      type: object
      properties:
        nonce: { type: string }
        message: { type: string }
      required: [nonce, message]
...
    UserResponse:
      type: object
      properties:
        user:
          allOf:
            - { $ref: "#/components/schemas/User" }
          nullable: true
      required: [user]
    ProfileUpdate:
      type: object
      additionalProperties: false
      properties:
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string }
        bio: { type: string }
...
    CreatorProfile:
      type: object
      additionalProperties: true
      properties:
        id: { type: string }
        userId: { type: string }
        username: { type: string }
        displayName: { type: string }
        isLive: { type: boolean }
      required: [id, userId, username, displayName]
    CreatorProfileInput:
      type: object
      properties:
        username: { type: string }
        displayName: { type: string }
        bio: { type: string }
        avatarUrl: { type: string }
        bannerUrl: { type: string }
        channelColor: { type: string }
        primaryCategoryId: { type: string }
      required: [username, displayName]
...